Do Bugs Foreshadow Vulnerabilities? RQ6 analysis
These scripts were used in the analysis of Apache httpd data set in the
context of the Do Bugs Foreshadow Vulnerabilities? study. The scripts do
assume independence and do not conform to the database schema that the
parent project may define. Instead, the scripts work off of a SQLite
database that was generated with data collected specifically for the
project.